drug/requirements.txt
flask==3.0.3
streamlit==1.41.1
pandas==2.2.3
numpy==2.2.1
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.32.3
huggingface_hub==0.27.1
scikit-learn==1.6.0
pypdf==5.1.0
